# api/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Tenant Search API
  description: Lightning-fast search API with simple and complex query support
  version: 1.0.0
  contact:
    name: Platform Team
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging-api.example.com/v1
    description: Staging
paths:
...
  /suggest:
    post:
      summary: Typeahead Suggestions
      description: Fast prefix-based suggestions for autocomplete
      operationId: suggest
      tags:
        - Search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestRequest'
      responses:
        '200':
          description: Suggestions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuggestResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuggestRequest:
      type: object
      required: [prefix]
      properties:
        prefix:
          type: string
          minLength: 1
          maxLength: 50
          example: "cust"
        entity:
          type: array
          items:
            type: string
          example: ["customer"]
        limit:
          type: integer
          minimum: 1
          maximum: 20
          default: 10
    SuggestResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              score:
                type: number
              context:
                type: object
                properties:
                  entity:
                    type: string
                  id:
                    type: string
        performance:
          type: object
          properties:
            took_ms:
              type: integer
            cached:
              type: boolean
...
security:
  - BearerAuth: []